feat: created modals for assigning user to camp
