+++ src/components/modals/RemoveUserFromCampModal.tsx
@@ -1,0 +1,52 @@
+// SecondModal.tsx
+
+import React from 'react'
+import {
+  Modal,
+  ModalOverlay,
+  ModalContent,
+  ModalHeader,
+  ModalFooter,
+  ModalBody,
+  ModalCloseButton,
+  Button,
+  Heading,
+  Text,
+} from '@chakra-ui/react'
+
+interface RemoveUserFromCampModalProps {
+  isOpen: boolean
+  onClose: () => void
+  onSubmit: () => void
+}
+
+const RemoveUserFromCampModal = ({
+  isOpen,
+  onClose,
+  onSubmit,
+}: RemoveUserFromCampModalProps) => {
+  // const handleSubmit = () =>{}
+
+  return (
+    <Modal isOpen={isOpen} onClose={onClose} size={'xs'}>
+      <ModalOverlay />
+      <ModalContent>
+        <ModalCloseButton />
+        <ModalBody>
+          <Heading>User</Heading>
+          <Text>Remove from Camp</Text>
+          {/* Add your second modal content here */}
+          <Text>Action will remove user from camp</Text>
+        </ModalBody>
+        <ModalFooter>
+          <Button colorScheme="blue" mr={3} onClick={onSubmit}>
+            Remove User
+          </Button>
+          <Button onClick={onClose}>Cancel</Button>
+        </ModalFooter>
+      </ModalContent>
+    </Modal>
+  )
+}
+
+export default RemoveUserFromCampModal
